Completely changed the direction of Wizzy (Major refactoring)
=====Wizzy 2D=====
Major decisions were made for Wizzy, the biggest one arguably being that most of the
3D Rendering was scrapped and a focus on 2D Rendering was attained. This will remain
the focus for Wizzy for a while and will be finely tuned. As of now, Wizzy has a
2D Batch Renderer which is, for the time being, fast enough. This Renderer can also
render to RenderTargets, where a null rendertarget means rendering directly to the
game window.

=====ResourceManagement=====
Once again, the ResourceManagement was completely refactored (5th time?). I am
however satisfied with this version so it will likely stay this way (albeit
some expansion). The new Resource Management System use a dynamic and serializable
"PropertyLibrary" instead of flags, which makes the loading and saving of resource
importing configurations way easier. ResourceManagement still uses meta files
the way this is handled makes way more sense now. Instead of using strings as
handles for resources, integers are now used. These are in the forms of structs
defined for each resource, meaning you cannot implicitly convert from for example
a Texture::Handle to a Shader::Handle. The API differs a bit but I refer you to
the top of ResourceManagement.cpp for a detailed description of how to use the new
API.
=====PropertyLibrary=====
PropertyLibrary class was implemented for beging able to dynamically store any of
the 4 basic data types and easily serialize/deserialize it.

=====Wizzy ECS=====
The ECS System used in Wizzy was previously one of my other library pdoes
(pure-data-oriented-ecs-system), but I decided to make a more specific Wizzy
implementation as there were some needs that were way easier to satisfy like
this. It's however heavily based on the original pdoes source code.

=====Misc=====
- Major improvements/maintenance in GL rendering implementations